refactoring the beneficiary redirect container into parts
